pyproject.toml
[project]
name = "fundamentos-bot"
version = "0.1.0"
description = "Bot for the programming fundamentals course at the University of Buenos Aires"
readme = "README.md"
authors = [
    { name = "Manuel Sanchez", email = "manuelsfdelavega@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "pydantic>=2.11.2",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
bot = "fundamentos_bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.2",
]
